Reconstructions
Pardon for any mistakens I've done in the past, as I didn't know how ASP.NET Core works, this time the project is reconstructed with properly ASP.NET Core (Razor Pages). For now, pages are not working, need to do some database connection and reconstruct the codebase. AGAIN, PLEASE DO NOT CONSIDER THIS AS FINAL VERSION, IT MIGHT SUBJECT TO CHANGE. DO NOT CONSIDER THIS AS THE FINAL PRODUCT