API is nearly complete, Frontend is in the work
Got the API, Used SwaggerUI to show some of the properties, eventhough I need to work more on them, like tell the UI to put stuff in the database. However things are almost done. WebUI is okay I guess, that's all for now because I was in a hurry. I will redesign the WebUI as soon as I can.
Need to reconstruct a lot of things, so that's all for now. Again, PLEASE DO NOT CONSIDER THIS WORK AS FINAL, IT IS SUBJECT TO CHANGE.